+++ src/debug/overlay/config/overlay_modules_config.js
@@ -1,5 +1,5 @@
 const OVERLAY_MODULES = {
-  echo: false,
+  echo: true,
   chancesDetail: false,
   state: true,
   chum: false,